improve UserContextTileLayer
